+++ asm2017/emu/src/disasm.c
@@ -159,7 +159,7 @@ uint disasm_pointer(memory_t *mem, uint64_t *ptr)
 	return memory_read(mem, *ptr - 2, 2);
 }
 
-uint8_t disasm_instr_length(uint id)
+uint disasm_instr_length(uint id)
 {
     return length[id];
 }
